+++ app/blueprints/vendor_blueprint.py
@@ -8,7 +8,6 @@ rating_blueprint = Blueprint('rating', __name__, url_prefix='{}/ratings'.format(
 
 vendor_controller = VendorController(request)
 
-'''VENDORS'''
 @vendor_blueprint.route('/', methods=['GET'])
 def list_vendors():
 	return vendor_controller.list_vendors()
@@ -31,56 +30,3 @@ def update_vendor(vendor_id):
 @Auth.has_permission('delete_vendor')
 def delete_vendor(vendor_id):
 	return vendor_controller.delete_vendor(vendor_id)
-
-
-'''VENDOR ENGAGEMENT'''
-@engagement_blueprint.route('/', methods=['GET'])
-def list_engagements():
-	return vendor_controller.list_vendor_engagements()
-
-@engagement_blueprint.route('/upcoming', methods=['GET'])
-def upcoming_engagements():
-	return vendor_controller.upcoming_vendor_engagements()
-
-@engagement_blueprint.route('/<int:engagement_id>', methods=['GET'])
-def get_engagements(engagement_id):
-	return vendor_controller.get_vendor_engagement(engagement_id)
-
-@engagement_blueprint.route('/', methods=['POST'])
-@Security.validator(['vendorId|required:int', 'startDate|required:date', 'endDate|required:date', 'status|int'])
-def create_engagement():
-	return vendor_controller.create_vendor_engagement()
-
-@engagement_blueprint.route('/<int:engagement_id>', methods=['PUT', 'PATCH'])
-@Security.validator(['vendorId|required:int', 'startDate|date', 'endDate|date', 'status|int', ])
-def update_engagement(engagement_id):
-	return vendor_controller.update_vendor_engagement(engagement_id)
-
-@engagement_blueprint.route('/<int:engagement_id>', methods=['DELETE'])
-@Auth.has_permission('delete_engagement')
-def delete_engagement(engagement_id):
-	return vendor_controller.delete_engagement(engagement_id)
-
-
-'''VENDOR RATING'''
-@rating_blueprint.route('/vendor/<int:vendor_id>', methods=['GET'])
-@Auth.has_permission('view_ratings')
-def list_ratings(vendor_id):
-	'''Gets all the ratings for a given vendor'''
-	return vendor_controller.list_ratings(vendor_id)
-
-
-@rating_blueprint.route('/<int:rating_id>', methods=['GET'])
-@Auth.has_permission('view_ratings')
-def get_vendor_rating(rating_id):
-	return vendor_controller.get_vendor_rating(rating_id)
-
-@rating_blueprint.route('/', methods=['POST'])
-@Security.validator(['vendor_id|required:int', 'rating|required:int'])
-def create_vendor_rating():
-	return vendor_controller.create_vendor_rating()
-
-@rating_blueprint.route('/<int:rating_id>', methods=['PUT', 'PATCH'])
-@Security.validator(['rating|int'])
-def update_rating(rating_id):
-	return vendor_controller.update_vendor_rating(rating_id)
